requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
Babel==2.6.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.6.1
docutils==0.14
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
imagesize==1.1.0
imutils==0.5.2
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.16
nose2==0.9.1
numpy==1.16.3
oauth2client==4.1.3
oauthlib==3.0.1
opencv-python==3.4.4.19
packaging==19.0
pep8==1.7.1
prettytable==0.7.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyAudio==0.2.11
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.0
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
pyzbar==0.1.8
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
RTIMULib==7.2.1
six==1.12.0
snowballstemmer==1.2.1
SpeechRecognition==3.8.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.3
typed-ast==1.3.4
uritemplate==3.0.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
